+++ packages/patternfly-4/react-charts/src/components/ChartBar/ChartBar.js
@@ -1,18 +1,35 @@
+/* eslint-disable react/require-default-props,react/prop-types */
 import React from 'react';
 import PropTypes from 'prop-types';
 import hoistNonReactStatics from 'hoist-non-react-statics';
 import { VictoryBar } from 'victory';
-import { default as ChartTheme } from '../ChartTheme/ChartTheme';
+import { getTheme } from '../ChartTheme/themes/theme-utils';
 
 export const propTypes = {
   /**
    * See TypeScript API docs: https://formidable.com/open-source/victory/docs/victory-bar/
    */
-  '': PropTypes.any
+  '': PropTypes.any,
+  /*
+   * Specifies the theme color; blue (default), green, or multi-color. Overridden by the theme property.
+   */
+  themeColor: PropTypes.string,
+  /*
+   * Specifies the theme variant; 'dark' or 'light' (default). Overridden by the theme property.
+   */
+  themeVariant: PropTypes.string
 };
 
+const ChartBar = ({
+  themeColor,
+  themeVariant,
+  theme = getTheme(themeColor, themeVariant), // destructure last
+  ...rest
+}) => (
+  <VictoryBar theme={theme} {...rest} />
+);
+
 // Note: VictoryBar.getDomain & VictoryBar.role must be hoisted
-const ChartBar = props => <VictoryBar theme={ChartTheme.default} {...props} />;
 hoistNonReactStatics(ChartBar, VictoryBar);
 ChartBar.propTypes = propTypes;