+++ packages/patternfly-4/react-charts/src/components/ChartLabel/ChartLabel.js
@@ -1,3 +1,4 @@
+/* eslint-disable react/require-default-props */
 import React from 'react';
 import PropTypes from 'prop-types';
 import hoistNonReactStatics from 'hoist-non-react-statics';
@@ -10,8 +11,11 @@ export const propTypes = {
   '': PropTypes.any
 };
 
+const ChartLabel = props => (
+  <VictoryLabel {...props} />
+);
+
 // Note: VictoryLabel.role must be hoisted
-const ChartLabel = props => <VictoryLabel {...props} />;
 hoistNonReactStatics(ChartLabel, VictoryLabel);
 ChartLabel.propTypes = propTypes;