+++ packages/patternfly-4/react-charts/src/components/ChartLegend/ChartLegend.d.ts
@@ -1,6 +1,10 @@
 import * as victory from 'victory';
 
-export interface ChartLegendProps extends victory.VictoryLegendProps {}
+export interface ChartLegendProps extends victory.VictoryLegendProps {
+  responsive?: boolean;
+  themeColor?: string;
+  themeVariant?: string;
+}
 
 declare const ChartLegend: React.ComponentClass<ChartLegendProps>;