+++ packages/patternfly-4/react-charts/src/components/ChartStack/ChartStack.d.ts
@@ -1,6 +1,9 @@
 import * as victory from 'victory';
 
-export interface ChartStackProps extends victory.VictoryStackProps {}
+export interface ChartStackProps extends victory.VictoryStackProps {
+  themeColor?: string;
+  themeVariant?: string;
+}
 
 declare const ChartStack: React.ComponentClass<ChartStackProps>;