+++ packages/patternfly-4/react-charts/src/components/ChartStack/ChartStack.js
@@ -1,22 +1,38 @@
+/* eslint-disable react/require-default-props,react/prop-types */
 import React from 'react';
 import PropTypes from 'prop-types';
 import hoistNonReactStatics from 'hoist-non-react-statics';
 import { VictoryStack } from 'victory';
-import { default as ChartTheme } from '../ChartTheme/ChartTheme';
+import { getTheme } from '../ChartTheme/themes/theme-utils';
 
 export const propTypes = {
   /**
    * See TypeScript API docs: https://formidable.com/open-source/victory/docs/victory-stack/
    */
-  '': PropTypes.any
+  '': PropTypes.any,
+  /*
+   * Specifies the theme color; blue (default), green, or multi-color. Overridden by the theme property.
+   */
+  themeColor: PropTypes.string,
+  /*
+   * Specifies the theme variant; 'dark' or 'light' (default). Overridden by the theme property.
+   */
+  themeVariant: PropTypes.string
 };
 
-// Note: VictoryStack.getChildren & VictoryStack.role must be hoisted
-const ChartStack = ({ children, ...props }) => (
-  <VictoryStack theme={ChartTheme.default} {...props}>
+const ChartStack = ({
+  children,
+  themeColor,
+  themeVariant,
+  theme = getTheme(themeColor, themeVariant), // destructure last
+  ...rest
+}) => (
+  <VictoryStack theme={theme} {...rest}>
     {children}
   </VictoryStack>
 );
+
+// Note: VictoryStack.getChildren & VictoryStack.role must be hoisted
 hoistNonReactStatics(ChartStack, VictoryStack);
 ChartStack.propTypes = propTypes;