+++ packages/patternfly-4/react-charts/src/components/ChartTheme/index.js
@@ -1,1 +1,2 @@
-export { default as ChartTheme } from './ChartTheme';
+export { ChartBaseTheme, ChartDonutTheme, ChartThemeColor, ChartThemeVariant } from './ChartTheme';
+export { getCustomTheme, getDonutTheme, getTheme } from './themes/theme-utils';